chore: organize components directory
