--- src/components/Button.module.scss
@@ -1,19 +1,0 @@
-.button {
-  padding: 0.5rem 1rem;
-  border: none;
-  border-radius: 8px;
-  background: var(--color-primary);
-  color: var(--color-text);
-  cursor: pointer;
-  transition: transform 0.2s ease, box-shadow 0.2s ease;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.3);
-}
-
-.button:hover {
-  transform: translateY(-2px) scale(1.03);
-  box-shadow: 0 6px 20px rgba(0, 0, 0, 0.5);
-}
-
-.button:active {
-  transform: scale(0.97);
-}
