--- src/components/EmptyState.module.scss
@@ -1,8 +1,0 @@
-.wrapper {
-  text-align: center;
-  padding: 2rem;
-}
-
-.actionButton {
-  margin-top: 1rem;
-}
