--- src/components/LoadingIndicator.module.scss
@@ -1,18 +1,0 @@
-@keyframes spin {
-  to { transform: rotate(360deg); }
-}
-
-.wrapper {
-  display: flex;
-  justify-content: center;
-  padding: 2rem;
-}
-
-.spinner {
-  width: 3rem;
-  height: 3rem;
-  border: 4px solid rgba(255, 255, 255, 0.3);
-  border-top-color: var(--color-primary);
-  border-radius: 50%;
-  animation: spin 1s linear infinite;
-}
