--- src/components/MetricComparison.module.scss
@@ -1,33 +1,0 @@
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-}
-
-.label {
-  font-weight: 600;
-  margin-bottom: 0.25rem;
-}
-
-.row {
-  display: flex;
-  align-items: baseline;
-  gap: 0.5rem;
-}
-
-.current {
-  font-size: 1.2rem;
-}
-
-.change {
-  font-size: 0.9rem;
-}
-
-.positive {
-  color: var(--color-primary);
-}
-
-.negative {
-  color: var(--color-negative);
-}
-
