--- src/components/NavBar.module.scss
@@ -1,12 +1,0 @@
-.nav {
-  background: var(--color-surface);
-
-  padding: 1rem;
-}
-
-.link {
-  color: var(--color-text);
-
-  margin-right: 1rem;
-  text-decoration: none;
-}
