--- src/components/TopBar.module.scss
@@ -1,43 +1,0 @@
-.bar {
-  position: sticky;
-  top: 0;
-  display: flex;
-  align-items: center;
-  padding: 0.5rem 1rem;
-  background: var(--color-surface);
-  border-bottom: 1px solid var(--color-border);
-
-  z-index: 1000;
-}
-
-.statContainer {
-  display: flex;
-}
-
-.statBadge {
-  background: var(--color-bg);
-
-  margin-right: 0.5rem;
-  padding: 0.25rem 0.5rem;
-  border-radius: 4px;
-  font-size: 0.875rem;
-}
-
-.actions {
-  margin-left: auto;
-  display: flex;
-  align-items: center;
-}
-
-.userArea {
-  display: flex;
-  align-items: center;
-  margin-left: 1rem;
-}
-
-.avatar {
-  width: 32px;
-  height: 32px;
-  border-radius: 50%;
-  margin-right: 0.5rem;
-}
