--- src/components/charts/ChartLegend.test.jsx
@@ -1,21 +1,0 @@
-import { describe, it, expect, vi } from 'vitest';
-import { render, fireEvent } from '@testing-library/react';
-import '@testing-library/jest-dom/vitest';
-import ChartLegend from './ChartLegend';
-
-const sampleItems = [
-  { label: 'Stocks', color: 'var(--color-accent-1)', visible: true },
-  { label: 'Bonds', color: 'var(--color-accent-2)', visible: false },
-];
-
-describe('ChartLegend', () => {
-  it('calls onToggle with label when an item is clicked', () => {
-    const handleToggle = vi.fn();
-    const { getByLabelText } = render(
-      <ChartLegend items={sampleItems} onToggle={handleToggle} />
-    );
-    const button = getByLabelText('toggle Stocks');
-    fireEvent.click(button);
-    expect(handleToggle).toHaveBeenCalledWith('Stocks');
-  });
-});
