+++ src/components/common/__tests__/LoadingIndicator.test.jsx
@@ -1,7 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom/vitest';
-import LoadingIndicator from './LoadingIndicator';
+import LoadingIndicator from '../LoadingIndicator';
 
 describe('LoadingIndicator', () => {
   it('renders spinner with status role', () => {
