+++ src/components/layout/__tests__/Sidebar.test.jsx
@@ -2,7 +2,7 @@ import { describe, it, expect } from 'vitest';
 import { render } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import '@testing-library/jest-dom/vitest';
-import Sidebar from './Sidebar';
+import Sidebar from '../Sidebar';
 
 describe('Sidebar', () => {
   it('renders default navigation items', () => {
